Add files via upload
Figure S2.  We visually confirmed that the palmar hyperlinearity in the original images was consistent with the hyperlinearity captured in the processed images and the generated hyperlinearity scores The 8 highest and lowest scoring images after Sato edge detection are shown for (A) thenar and (B) palmar images. Comparison to original photos (not shown for consent reasons) confirmed the generated scores and processed images were appropriate.